Declare missing dependency for libcuda.so
Testing shows that libcuda.so depends on its sibling (located in the
same package) libnvcucompat.so, although this is not visible in the
dynamic elf section. This dependency is probably coming from a dynamic
call to `dlopen`. This commit explicitly adds this dependency via
patchelf to avoid relying on LD_LIBRARY_PATH or another global, dynamic
mechanism, and instead relies on Nix standard patching magic.